api: moved log-related tests to api/log
